added BoardOverlay object to chat.py, moved BoardFSM -> fsm.py, renamed FSM
